+++ package.json
@@ -60,8 +60,8 @@
     "vuex-persistedstate": "^4.0.0"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
     "@acedatacloud/transmart-cli": "^0.5.0",
+    "@rollup/plugin-replace": "^3.0.0",
     "@types/lodash": "^4.14.172",
     "@types/markdown-it": "^12.2.3",
     "@types/node": "^16.7.4",
@@ -92,7 +92,7 @@
     "tailwindcss": "^3.4.0",
     "typescript": "^4.9.5",
     "vite": "^2.9.16",
-    "vite-bundle-visualizer": "^1.0.0",
+    "vite-bundle-visualizer": "^1.2.1",
     "vue-tsc": "^1.2.0"
   },
   "bugs": {
