+++ src/utils/index.ts
@@ -1,3 +1,6 @@
 export * from './baseUrl';
 export * from './mode';
 export * from './log';
+export * from './initializer';
+export * from './highlight';
+export * from './is';
